pyproject.toml
[project]
name = "langg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.4",
    "langchain-tavily>=0.2.5",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
